+++ src/js/gutenberg-overrides/packages/block-library/CHANGELOG.md
@@ -6,7 +6,9 @@
 
 - 'cover-image' to 'cover' block ([packages/block-libary/build-module/cover](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-libary/build-module/cover))
 
-- 'gallery' block ([packages/block-libary/build-module/gallery/edit.js](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-libary/build-module/gallery/gallery-image.js)) and ([packages/block-libary/build-module/gallery/edit.js](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-libary/build-module/gallery/gallery-image.js))
+- 'gallery' block ([packages/block-libary/build-module/gallery/edit.js](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-libary/build-module/gallery/edit.js)) and ([packages/block-libary/build-module/gallery/edit.js](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-libary/build-module/gallery/gallery-image.js))
+
+- 'image' block ([packages/block-libary/build-module/image/edit.js](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-libary/build-module/image/edit.js))
 
 ## 2.0.0 2018-08